students can be listed and added to specific class

+++ src/routes/users.ts
@@ -1,16 +1,68 @@
-import Joi from '@hapi/joi';
-import bcrypt from 'bcrypt';
-import express, { Router } from 'express';
+import Joi from "@hapi/joi";
+import bcrypt from "bcrypt";
+import express, { Router } from "express";
 
-import { authorize } from '../middlewares/authorize';
-import { Role } from '../models/enums';
-import { createUser, updateUser, getUsersByType, deleteUser, findUserByID, addStudentInParent, getStudentsByParent } from '../services/users';
-import { wrapAsync } from '../utils/asyncHandler';
+import { authorize } from "../middlewares/authorize";
+import { Role } from "../models/enums";
+import {
+  createUser,
+  updateUser,
+  getUsersByType,
+  deleteUser,
+  findUserByID,
+  addStudentInParent,
+  getStudentsByParent,
+  getAllByTypeAndKeyword,
+} from "../services/users";
+import { wrapAsync } from "../utils/asyncHandler";
 
-import { Request, isUserReq } from './interfaces';
+import { Request, isUserReq } from "./interfaces";
 
 const router = Router();
 
+//swager documentation needs to be added here
+
+router.get(
+  "/users/:role/byKeyword",
+  authorize([Role.ADMIN]),
+  wrapAsync(async (req: Request, res: express.Response) => {
+    if (!isUserReq(req)) {
+      throw new Error("User not found in session");
+    }
+    const { limit, offset, role, keyword } = await Joi.object({
+      offset: Joi.number().integer().default(0).failover(0).label("Offset"),
+      limit: Joi.number().integer().default(10).failover(10).label("Limit"),
+      role:
+        req.user.role === Role.SUPER_ADMIN
+          ? Joi.string().valid(Role.ADMIN).required().label("User role")
+          : Joi.string()
+              .valid(Role.TEACHER, Role.PARENT, Role.STUDENT)
+              .required()
+              .label("User role"),
+      keyword: Joi.string().allow("", null).default("").label("Keyword"),
+    }).validateAsync({
+      offset: req.query.offset,
+      limit: req.query.limit,
+      keyword: req.query.keyword,
+      role: req.params.role,
+    });
+    const [users, total] = await getAllByTypeAndKeyword(
+      role,
+      offset,
+      limit,
+      keyword
+    );
+
+    const dataParsed = users.map((user) => ({
+      id: user.id,
+      name: user.firstName + " " + user.lastName,
+    }));
+    res.send({
+      total,
+      data: dataParsed,
+    });
+  })
+);
 /**
  * @swagger
  * /users:
@@ -58,36 +110,43 @@ const router = Router();
  *       500:
  *         $ref: '#/components/responses/InternalError'
  */
-router.get('/users/:role', authorize([Role.ADMIN]), wrapAsync(async (req: Request, res: express.Response) => {
+router.get(
+  "/users/:role",
+  authorize([Role.ADMIN]),
+  wrapAsync(async (req: Request, res: express.Response) => {
     if (!isUserReq(req)) {
-        throw new Error('User not found in session');
+      throw new Error("User not found in session");
     }
-
-    const { limit, offset, role } = await Joi
-        .object({
-            offset: Joi.number().integer().default(0).failover(0).label('Offset'),
-            limit: Joi.number().integer().default(10).failover(10).label('Limit'),
-            role: req.user.role === Role.SUPER_ADMIN ? Joi.string().valid(Role.ADMIN).required().label('User role') : Joi.string().valid(Role.TEACHER, Role.PARENT).required().label('User role'),
-        })
-        .validateAsync({
-            offset: req.query.offset,
-            limit: req.query.limit,
-            role: req.params.role,
-        });
-
+    const { limit, offset, role } = await Joi.object({
+      offset: Joi.number().integer().default(0).failover(0).label("Offset"),
+      limit: Joi.number().integer().default(10).failover(10).label("Limit"),
+      role:
+        req.user.role === Role.SUPER_ADMIN
+          ? Joi.string().valid(Role.ADMIN).required().label("User role")
+          : Joi.string()
+              .valid(Role.TEACHER, Role.PARENT, Role.STUDENT)
+              .required()
+              .label("User role"),
+    }).validateAsync({
+      offset: req.query.offset,
+      limit: req.query.limit,
+      role: req.params.role,
+    });
     const [users, total] = await getUsersByType(role, offset, limit);
 
     res.send({
-        total,
-        data: users.map((user) => ({
-            id: user.id,
-            email: user.email,
-            firstName: user.firstName,
-            lastName: user.lastName,
-            role: user.role,
-        })),
+      total,
+      data: users.map((user) => ({
+        id: user.id,
+        email: user.email,
+        firstName: user.firstName,
+        lastName: user.lastName,
+        role: user.role,
+        dob: user.DOB,
+      })),
     });
-}));
+  })
+);
 
 /**
  * @swagger
@@ -119,19 +178,23 @@ router.get('/users/:role', authorize([Role.ADMIN]), wrapAsync(async (req: Reques
  *       500:
  *         $ref: '#/components/responses/InternalError'
  */
-router.get('/users/me', authorize(), wrapAsync(async (req: Request, res: express.Response) => {
+router.get(
+  "/users/me",
+  authorize(),
+  wrapAsync(async (req: Request, res: express.Response) => {
     if (!isUserReq(req)) {
-        throw new Error('User not found in session');
+      throw new Error("User not found in session");
     }
 
     res.send({
-        id: req.user.id,
-        email: req.user.email,
-        firstName: req.user.firstName,
-        lastName: req.user.lastName,
-        role: req.user.role,
+      id: req.user.id,
+      email: req.user.email,
+      firstName: req.user.firstName,
+      lastName: req.user.lastName,
+      role: req.user.role,
     });
-}));
+  })
+);
 
 /**
  * @swagger
@@ -181,28 +244,40 @@ router.get('/users/me', authorize(), wrapAsync(async (req: Request, res: express
  *       500:
  *         $ref: '#/components/responses/InternalError'
  */
-router.put('/users/me', authorize(), wrapAsync(async (req: Request, res: express.Response) => {
+router.put(
+  "/users/me",
+  authorize(),
+  wrapAsync(async (req: Request, res: express.Response) => {
     if (!isUserReq(req)) {
-        throw new Error('User not found in session');
+      throw new Error("User not found in session");
     }
 
-    const user = await Joi
-        .object({
-            firstName: Joi.string().trim().min(1).max(50).required().label('First name'),
-            lastName: Joi.string().trim().min(1).max(50).required().label('Last name'),
-        })
-        .validateAsync(req.body);
+    const user = await Joi.object({
+      firstName: Joi.string()
+        .trim()
+        .min(1)
+        .max(50)
+        .required()
+        .label("First name"),
+      lastName: Joi.string()
+        .trim()
+        .min(1)
+        .max(50)
+        .required()
+        .label("Last name"),
+    }).validateAsync(req.body);
 
     const _user = await updateUser(req.user.id, user);
 
     res.send({
-        id: _user.id,
-        email: req.user.email,
-        firstName: _user.firstName,
-        lastName: _user.lastName,
-        role: req.user.role,
+      id: _user.id,
+      email: req.user.email,
+      firstName: _user.firstName,
+      lastName: _user.lastName,
+      role: req.user.role,
     });
-}));
+  })
+);
 
 /**
  * @swagger
@@ -267,30 +342,54 @@ router.put('/users/me', authorize(), wrapAsync(async (req: Request, res: express
  *       500:
  *         $ref: '#/components/responses/InternalError'
  */
-router.post('/users', authorize([Role.SUPER_ADMIN, Role.ADMIN]), wrapAsync(async (req: Request, res: express.Response) => {
+router.post(
+  "/users",
+  authorize([Role.SUPER_ADMIN, Role.ADMIN]),
+  wrapAsync(async (req: Request, res: express.Response) => {
     if (!isUserReq(req)) {
-        throw new Error('User not found in session');
+      throw new Error("User not found in session");
     }
 
-    const { firstName, lastName, email, password, role } = await Joi
-        .object({
-            firstName: Joi.string().trim().min(1).max(50).required().label('First name'),
-            lastName: Joi.string().trim().min(1).max(50).required().label('Last name'),
-            email: Joi.string().trim().lowercase().email().required().label('Email'),
-            password: Joi.string().trim().min(1).max(50).required().label('Password'),
-            role: req.user.role === Role.SUPER_ADMIN ? Joi.string().valid(Role.ADMIN).required().label('User role') : Joi.string().valid(Role.TEACHER, Role.PARENT).required().label('User role'),
-        })
-        .validateAsync(req.body);
+    const { firstName, lastName, email, password, role } = await Joi.object({
+      firstName: Joi.string()
+        .trim()
+        .min(1)
+        .max(50)
+        .required()
+        .label("First name"),
+      lastName: Joi.string()
+        .trim()
+        .min(1)
+        .max(50)
+        .required()
+        .label("Last name"),
+      email: Joi.string().trim().lowercase().email().required().label("Email"),
+      password: Joi.string().trim().min(1).max(50).required().label("Password"),
+      role:
+        req.user.role === Role.SUPER_ADMIN
+          ? Joi.string().valid(Role.ADMIN).required().label("User role")
+          : Joi.string()
+              .valid(Role.TEACHER, Role.PARENT)
+              .required()
+              .label("User role"),
+    }).validateAsync(req.body);
 
     const hashedPassword = await bcrypt.hash(password, 8);
-    const user = await createUser(firstName, lastName, email, hashedPassword, role);
+    const user = await createUser(
+      firstName,
+      lastName,
+      email,
+      hashedPassword,
+      role
+    );
 
     res.send({
-        id: user.id,
-        email: user.email,
-        role: user.role,
+      id: user.id,
+      email: user.email,
+      role: user.role,
     });
-}));
+  })
+);
 
 /**
  * @swagger
@@ -318,23 +417,25 @@ router.post('/users', authorize([Role.SUPER_ADMIN, Role.ADMIN]), wrapAsync(async
  *       500:
  *         $ref: '#/components/responses/InternalError'
  */
-router.delete('/users/:userId', authorize([Role.ADMIN]), wrapAsync(async (req: Request, res: express.Response) => {
+router.delete(
+  "/users/:userId",
+  authorize([Role.ADMIN]),
+  wrapAsync(async (req: Request, res: express.Response) => {
     if (!isUserReq(req)) {
-        throw new Error('User not found in session');
+      throw new Error("User not found in session");
     }
 
-    const { userId } = await Joi
-        .object({
-            userId: Joi.string().uuid().required().label('User ID'),
-        })
-        .validateAsync({
-            userId: req.params.userId,
-        });
+    const { userId } = await Joi.object({
+      userId: Joi.string().uuid().required().label("User ID"),
+    }).validateAsync({
+      userId: req.params.userId,
+    });
 
     await deleteUser(userId);
 
     res.status(204).send();
-}));
+  })
+);
 
 /**
  * @swagger
@@ -384,97 +485,142 @@ router.delete('/users/:userId', authorize([Role.ADMIN]), wrapAsync(async (req: R
  *       500:
  *         $ref: '#/components/responses/InternalError'
  */
-router.put('/users/:userId', authorize(), wrapAsync(async (req: Request, res: express.Response) => {
+router.put(
+  "/users/:userId",
+  authorize(),
+  wrapAsync(async (req: Request, res: express.Response) => {
     if (!isUserReq(req)) {
-        throw new Error('User not found in session');
+      throw new Error("User not found in session");
     }
 
-    const { userId, firstName, lastName, email } = await Joi
-        .object({
-            userId: Joi.string().uuid().required().label('User ID'),
-            firstName: Joi.string().trim().min(1).max(50).required().label('First name'),
-            lastName: Joi.string().trim().min(1).max(50).required().label('Last name'),
-            email: Joi.string().trim().lowercase().email().required().label('Email'),
-        })
-        .validateAsync({
-            userId: req.params.userId,
-            ...req.body,
-        });
+    const { userId, firstName, lastName, email } = await Joi.object({
+      userId: Joi.string().uuid().required().label("User ID"),
+      firstName: Joi.string()
+        .trim()
+        .min(1)
+        .max(50)
+        .required()
+        .label("First name"),
+      lastName: Joi.string()
+        .trim()
+        .min(1)
+        .max(50)
+        .required()
+        .label("Last name"),
+      email: Joi.string().trim().lowercase().email().required().label("Email"),
+    }).validateAsync({
+      userId: req.params.userId,
+      ...req.body,
+    });
 
     const _user = await updateUser(userId, { firstName, lastName, email });
 
     res.send({
-        id: _user.id,
-        email: req.user.email,
-        firstName: _user.firstName,
-        lastName: _user.lastName,
-        role: req.user.role,
+      id: _user.id,
+      email: req.user.email,
+      firstName: _user.firstName,
+      lastName: _user.lastName,
+      role: req.user.role,
     });
-}));
+  })
+);
 
-router.post('/students/parent/:parentId', authorize([Role.SUPER_ADMIN, Role.ADMIN]), wrapAsync(async (req: Request, res: express.Response) => {
+router.post(
+  "/students/parent/:parentId",
+  authorize([Role.SUPER_ADMIN, Role.ADMIN]),
+  wrapAsync(async (req: Request, res: express.Response) => {
     if (!isUserReq(req)) {
-        throw new Error('User not found in session');
+      throw new Error("User not found in session");
     }
 
-    const { firstName, lastName, email, password, role, parentId } = await Joi
-        .object({
-            firstName: Joi.string().trim().min(1).max(50).required().label('First name'),
-            lastName: Joi.string().trim().min(1).max(50).required().label('Last name'),
-            email: Joi.string().trim().lowercase().email().required().label('Email'),
-            password: Joi.string().trim().min(1).max(50).required().label('Password'),
-            role: req.user.role === Role.SUPER_ADMIN ? Joi.string().valid(Role.ADMIN).required().label('User role') : Joi.string().valid(Role.TEACHER, Role.PARENT, Role.STUDENT).required().label('User role'),
-            parentId: Joi.string().uuid().required().label('Parent ID'),
-        })
-        .validateAsync({
-            ...req.body,
-            parentId: req.params.parentId,
-        });
+    const {
+      firstName,
+      lastName,
+      email,
+      password,
+      role,
+      parentId,
+    } = await Joi.object({
+      firstName: Joi.string()
+        .trim()
+        .min(1)
+        .max(50)
+        .required()
+        .label("First name"),
+      lastName: Joi.string()
+        .trim()
+        .min(1)
+        .max(50)
+        .required()
+        .label("Last name"),
+      email: Joi.string().trim().lowercase().email().required().label("Email"),
+      password: Joi.string().trim().min(1).max(50).required().label("Password"),
+      role:
+        req.user.role === Role.SUPER_ADMIN
+          ? Joi.string().valid(Role.ADMIN).required().label("User role")
+          : Joi.string()
+              .valid(Role.TEACHER, Role.PARENT, Role.STUDENT)
+              .required()
+              .label("User role"),
+      parentId: Joi.string().uuid().required().label("Parent ID"),
+    }).validateAsync({
+      ...req.body,
+      parentId: req.params.parentId,
+    });
 
     const hashedPassword = await bcrypt.hash(password, 8);
 
     const parent = await findUserByID(parentId);
 
-    const student = await createUser(firstName, lastName, email, hashedPassword, role);
+    const student = await createUser(
+      firstName,
+      lastName,
+      email,
+      hashedPassword,
+      role
+    );
 
     await addStudentInParent(student, parent);
 
     res.send({
-        id: student.id,
-        email: student.email,
-        role: student.role,
+      id: student.id,
+      email: student.email,
+      role: student.role,
     });
-}));
+  })
+);
 
-router.get('/students/parent/:parentId', authorize([Role.SUPER_ADMIN, Role.ADMIN]), wrapAsync(async (req: Request, res: express.Response) => {
+router.get(
+  "/students/parent/:parentId",
+  authorize([Role.SUPER_ADMIN, Role.ADMIN]),
+  wrapAsync(async (req: Request, res: express.Response) => {
     if (!isUserReq(req)) {
-        throw new Error('User not found in session');
+      throw new Error("User not found in session");
     }
 
-    const { offset, limit, parentId } = await Joi
-        .object({
-            offset: Joi.number().integer().default(0).failover(0).label('Offset'),
-            limit: Joi.number().integer().default(10).failover(10).label('Limit'),
-            parentId: Joi.string().uuid().required().label('Parent ID'),
-        })
-        .validateAsync({
-            offset: req.query.offset,
-            limit: req.query.limit,
-            parentId: req.params.parentId,
-        });
+    const { offset, limit, parentId } = await Joi.object({
+      offset: Joi.number().integer().default(0).failover(0).label("Offset"),
+      limit: Joi.number().integer().default(10).failover(10).label("Limit"),
+      parentId: Joi.string().uuid().required().label("Parent ID"),
+    }).validateAsync({
+      offset: req.query.offset,
+      limit: req.query.limit,
+      parentId: req.params.parentId,
+    });
 
     const [users, total] = await getStudentsByParent(parentId, offset, limit);
 
     res.send({
-        total,
-        data: users.map((user) => ({
-            id: user.id,
-            email: user.email,
-            firstName: user.firstName,
-            lastName: user.lastName,
-            role: user.role,
-        })),
+      total,
+      data: users.map((user) => ({
+        id: user.id,
+        email: user.email,
+        firstName: user.firstName,
+        lastName: user.lastName,
+        role: user.role,
+      })),
     });
-}));
+  })
+);
 
 export default router;
